redux: Use console when logging performance.
For detailed performance investigations, this is much more convenient
than the in-app `timing` log.  In particular:

 * It makes it possible to watch the log while using the app, without
   disrupting the use of the app to navigate to another screen.

 * It makes it possible to read longer stretches of logs, using a
   terminal, and to copy-paste from them for discussion.

 * It makes it possible to add other types of log entries in an easy
   unstructured way, as text, without having to build UI for each one.

The in-app log is still potentially useful for logging that we might
want to ask a user who isn't doing development on the app to send us a
screenshot of.  In particular, we currently use it for logging how
long it takes to rehydrate.  (The answer is about 1600ms, on a
Pixel 2 XL when logged into an account on czo with lots of unreads.)

But that doesn't apply to data like this, which is too voluminous for
us to leave on in normal release builds, and too voluminous to
reasonably get the data from our little UI for the in-app log anyway.