wip try to fix types in UnreadCards
It turns out this is blocked by the type of SectionList upstream
lacking a $ReadOnly on a particular piece, namely SectionBase which is
the type of the data describing one section.

That's a real simple fix (once diagnosed), so I may see about getting
it fixed upstream.  Until it is, probably the best thing we can do
here is to keep the fixme and just add a comment pointing at the
upstream issue/PR I file.

Well, that and maybe some of these other changes adding $ReadOnly and
giving names to more types.