wip android build: Update to recent Kotlin and bleeding-edge Android Gradle Plugin.
Probably shouldn't merge this until it can point to an AGP plugin
that's not alpha.

But the results are pretty great!  As a pure test of Gradle
configuration time: `./gradlew help` goes from ~5.1s to ~0.8s.
And most practically: `tools/test android` on no changes goes
from ~11.3s... to about 2.2s, a speedup of over 5x.  Sweet.

See here and the linked gradle/gradle tracking task:
  https://blog.gradle.org/introducing-configuration-caching#requirements-and-limitations