compose: In prompt for PM, drop ill-formed "@" sign.
When you open the compose box for a 1:1 PM, we've been showing a
placeholder like "Message @Greg Price".

This doesn't make a lot of sense.  It's not a valid @-mention
(that'd be "@**Greg Price**", with double-stars); and it has that
space in it (as long as the user's full_name has a space, which
in many languages they almost always do), which makes it look like
the "@", whatever it's doing, would only apply to e.g. "Greg" and
not "Greg Price".

Instead, do what the webapp does: use the person's name naturally,
and drop the "@".

From the history (#594/#599), it looks like the original idea here
was to ape Slack, which apparently prompts with something like
"Message @gregprice".  But that works because it's using a username;
there's no space, and in fact that's the normal syntax for an
@-mention in Slack.  Zulip doesn't have usernames, and a user's
natural full name is a bad fit for just substituting into the
same pattern.