users [nfc]: Drop a tiny, unhelpful optimization in filterUserList.
When the list is empty, that's exactly when there's nothing
significant to be gained by returning the identical list rather than
proceeding with the `Array#filter`.

In fact there's really nothing at all to gain: the one way it might
help is by enabling an "identical, therefore unchanged" optimization
somewhere downstream, but at the one call site, we never pass the
result as a React prop or anywhere else that might try to apply that
optimization.

Removing it is helpful because it lets us mark this function's input
as a read-only array while keeping the output read-write, effectively
signalling that this function returns a new copy rather than the
original array.  We want the output read-write so that the caller can
go on to `Array#sort` it (without making an additional, unnecessary
copy); and we'll want the input read-only shortly so that we can make
it take `UserOrBot` elements and stay compatible with being passed an
array that's supposed to contain specifically `User`.