narrow: In isNarrowValid use only user IDs, not emails, for PMs.
This check is itself the last remaining place where we consume
the emails from a PM narrow and then try to find the users in our
data structures; everything else relies on the user IDs.  So there
are no longer any potential crashes that this check could prevent.

We're near completion on removing those emails entirely, and in
particular we're about to make them emails invisible to code
outside the `narrow` module.  That will help ensure we don't add
more such places in the future -- and it requires removing this
now-superfluous check.