wip narrowsSelectors: Throw, not return bogus data, on unknown stream.
TODO fact-check this more carefully

This selector is called only from parts of the UI that we should only
get to if the narrow is indeed a stream or topic narrow, and one where
we know about the stream.  So just throw an error if those somehow
fail to be true.