eslint: Disable no-control-regex.
I don't really understand why someone ever thought this rule was a
good idea.  The upstream doc for it:
  https://eslint.org/docs/rules/no-control-regex
says:

> These characters are rarely used in JavaScript strings so a regular
> expression containing these characters is most likely a mistake.

That would almost make sense to me as an explanation (though one I'd
still disagree with) for a rule that applied to string literals in
general.  But it has nothing to do with regexps -- if anything, a
control character in a regexp is disproportionately likely to be there
precisely to filter out or reject such characters in input data.