store [nfc]: Make `customReplacer` and `customReviver` non-arrow functions.
This is only NFC because we haven't yet been using `this` in
`customReplacer` or `customReviver`.

The `replacer` function that's passed to `JSON.stringify`, or
`jsan.stringify`, is called with an interesting value as `this`: the
"holder" object, as JSON.stringify walks over the entire object. Or,
at least, it does so as long as the `replacer` function is a
non-arrow function. For an arrow function, `this` is set based on
the lexical scope in which it was defined.

remotedev-serialize has always taken care to pass a non-arrow
function in the one place it calls `jsan.stringify`. But then it
calls our code, `customReplacer` -- and we get to decide whether we
want `customReplacer`'s `this` to have that interesting value. So
far, we've decided we don't, by making `customReplacer` an arrow
function. If we make it a non-arrow function, its `this` will be the
`this` of its caller.

That decision doesn't seem well-considered; and, in particular,
we'll start using that interesting value in an upcoming commit.

The story is very similar for the `reviver` passed to
`JSON.stringify` / `jsan.stringify`.