wip EXP try converting presence state -- but argh, server sends emails
And in particular, converting at *initial load* is going to be
expensive -- at a minimum we have to make a pass once through
the list of users in order to build the map to find user ID by email.

Maybe leave the state as it is and just have a caching selector?