narrow [nfc]: Inline caseNarrow in keyFromNarrow.
This provides another demo of the style we might convert our
Narrow consumers to across the codebase, now that we have a
reasonable data structure for it.

More immediately, it will allow us to drop emails from the caseNarrow
interface, while still for now using them in apiNarrowOfNarrow and
therefore keeping them around in the Narrow values.