narrow [nfc]: Inline caseNarrow in apiNarrowOfNarrow.
This provides a demo of the style we might convert our Narrow
consumers to across the codebase, now that we have a reasonable
data structure for it.

More immediately, we want to drop emails from the caseNarrow
interface, to minimize the amount of code that's still exposed
to emails for PMs... but for now we still need to use them here,
until we drop support for pre-2.1 servers.  This change
decouples those.

Also add a TODO comment that we might be able to find by searching
when we drop support for those old servers.