ChatScreen: Use React Hooks.
In the upcoming React Navigation 5 upgrade, the
`withNavigationFocus` HOC is removed [1], and the `useIsFocus` and
`useFocusEffect` hooks are introduced. So, get ready to use those
hooks by making `ChatScreen` a hooks-based component in the first
place. React recommends gradually adopting Hooks, one component at a
time, as needed [2].

Since this is the first use of Hooks in our own code, we enable the
ESLint config [3] recommended [4] by React. We had already installed
it.

The most noticeable difference is that `ChatScreen` is a plain
function, not an ES6 class.

The return value of that function is meant to mirror the return
value of the `render` method, so the JSX stays pretty much the same.

Translate the `componentDidMount` and `componentDidUpdate` logic
into a custom hook, `useFetchMessages`. Address one hiccup in that
translation by using a few separate `useEffect`s [5]: it's not
ergonomic to compare an updated prop to its previous value, like
`prevProps` vs. `this.props` in `componentDidUpdate`. (There is a
way to do it, using `useRef`, but we don't do that because it looks
very awkward [6].)

See also discussion [7] about some small functional changes, like
how `componentDidMount` runs before the UI has been updated, while
`useEffect` runs after. We do expect this commit to be close to NFC,
though.

[1] https://reactnavigation.org/docs/upgrading-from-4.x/#higher-order-components.
    It's still available in the @react-navigation/compat package,
    though.

[2] https://reactjs.org/docs/hooks-intro.html#gradual-adoption-strategy

[3] https://www.npmjs.com/package/eslint-plugin-react-hooks

[4] https://reactjs.org/docs/hooks-reference.html#conditionally-firing-an-effect

[5] https://reactjs.org/docs/hooks-effect.html#tip-use-multiple-effects-to-separate-concerns

[6] https://reactjs.org/docs/hooks-faq.html#how-to-get-the-previous-props-or-state

[7] https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/converting.20to.20Hooks/near/1092349