unread [nfc]: Rename an old "direct" selector as being legacy.
We're going to switch the data structure we use for this to
something that's efficient to keep updated.  The new one will be
just as well suited for code that consumes this data structure as
the old -- in fact better for some of them -- so we'll have the
consumers switch to using the new data structure, too.  That means
the selector that provides the old data structure is going away.

As a first step, we rename the old selector to a "legacy" name,
making room for a selector providing the new data structure to
take its name.