wip and convert unread.streams to Immutable!
TODO:
  turn that LOGIN_SUCCESS thing into an issue
  do a bit of perf measurement
  later: start expanding state to take over consumer selectors' jobs, too

As we convert the other parts of the unread state, I'll want to fold
their reducers into this file too, and in fact combine the logic.  No
need to have four copies of all this, most of which is the same.

Probably as we do so just make this new reducer operate on the whole
subtree.  For now it can leave alone the parts of it that other
reducers have handled.

While making this conversion, I notice that this reducer doesn't reset
state on LOGIN_SUCCESS like it does for LOGOUT and ACCOUNT_SWITCH.  In
general we're pretty consistent about resetting state on those latter
two, but many of our reducers do so on LOGIN_SUCCESS while many others
don't.  We should almost surely be consistent; but I'm not entirely
confident which way is the right one, or if there could be regressions
from simply switching if not careful about it, so for now just add a
TODO comment and leave fixing that for a separate effort.