wip types [nfc]: Test BoundedDiff; fix some details in its jsdoc.
TODO update jsdoc re the Flow 6225 thing

Turns out the resulting type is actually exact, not inexact!
That's behavior inherited from $Diff.

Not sure if that changed at some point since I originally wrote this,
or if it was always this way -- could certainly be the latter, as
after all this didn't have tests. ;-)

Also don't try to characterize the behavior if the arguments aren't
read-only.