caughtUp: Fix bug that overwrote CaughtUp state.
When we made a request for a narrow, we would overwrite the CaughtUp
state for that narrow in the other direction - i.e. if we were fetching
old messages, then we would forget that we've seen the most recent
message.

This commit fixes that by only updating each bit of the CaughtUp state
from false to true (on a fetch-messages result), and never the other
way around.