webview [nfc]: Break up conditionals in long-press handler.
This saves us from having several different conditional chains that
rely on being parallel to each other, and lets us instead handle each
type of user intent in a single place.

The precedence here of the `a` case and the `.header` case is chosen
to match the behavior of `handleOutboundEvents.js` when it gets the
event/message we emit here: if `href` is non-null, then we act on that
without checking whether `target` is `link` or something else.  So if
we were to have an element matching `a.header`, this change would
alter the `target` property of the outbound message we emit to the
outside-webview code, but it preserves the actual behavior.