webview [nfc]: Get message ID directly from element in long-press handler.
At this point we've already looked for an enclosing message element;
instead of looking again with our own loop, just get data-msg-id
directly from that.

Moreover, if there is no enclosing message element, then just do
nothing rather than emit an event with a bogus "message ID", as
`getMessageIdFromNode` would have us do.  The end result is the
same -- the handler in `handleOutboundEvents` just ignores this
type of event if it has a bogus message ID -- but it's cleaner
not to take it that far in the first place.