webview js [nfc]: Rename some "node" locals that are actually Element.
Node and Element are two different types in the DOM; Node is the
base class of Element, and Element is where many of the methods we
use live.  Where the type is actually Element but the name has been
saying "Node", reduce confusion by either saying "element" or
letting the type speak for itself.