wip webview html [nfc]: Cut data-msg-id on date separators.
These inherently don't make a ton of sense.

When first introduced, we did actually use them -- specifically in
the original algorithm for identifying what messages are in view
and therefore should be marked as read.

But that algorithm was pretty broken for several reasons, which is
why we replaced it with a new one in 7f968e568, back in 2018.  And
in fact one of those bugs was directly related to its use of these
message IDs on date separators:

  > * When the bottom of the screen has a date separator, after which
  >   offscreen is a recipient bar, after which further offscreen is a
  >   message... we'll treat seeing that date separator as if it meant
  >   seeing the message.

I believe these have been unused since then.  Certainly I don't see
anything in the existing code now that would use them.