tools/test: Use stdout consistently.
These are errors of a sort, yes, but specifically they're test
failures -- which makes them part of the normal output of this
program, a test script.

We still use stderr in one place, for printing a usage message, which
we do when there's an error in the invocation of the script itself.