wip unread [nfc]: Fix a couple of parameter names.
These are not streams.  Each of them is a bundle of data with things
like a set of unread topics, with the number of unread messages in
each topic.

They are distinguished here, though, relative to the other data
floating around, by being the set of data we're working with for a
particular stream, as opposed to the whole dataset on the one hand or
a more-specific piece of data on the other.  In other words, they're
the per-stream blob of data.