api errors: Throw an exception to reach logging code.
This is NFC in the case where we reach `makeErrorFromApi`
and throw an exception it returns.

If there's an exception somewhere else -- which would be a bug in
this code -- then this change means that we now reach the logging
code that got moved inside this new `catch` block.  That seems
all to the good.  The `httpStatus` and `data` values are likely
to be undefined.