WIP drop optionalness on isFooNarrow predicates
Flow finds no problem, except in tests for this feature.

Still, could be a rogue caller somewhere.  And there's not really
much upside to this cleanup.  Probably best to leave it this way
until these get dissolved entirely by making Narrow non-opaque.
Converting callsites over to direct inspection will be a bigger
win and a perfect opportunity to audit that they don't seem to be
relying on passing void.