wip flow: Cut superseded "yarn test:flow-todo"
TODO add lint that all @flow is strict mode
  and/or, move these lints to [lints] section

The `no-weak-types` rule is redundant with the `unclear-type`
flowlint, which we have enabled in strict mode.  And strict mode,
in turn, is on for nearly all our files that have Flow at all.

The rules for annotating parameter and return types are redundant,
as far as we actually want them, with what Flow now always requires
in order to make types-first mode work.  Conversely, when we have
helpers local to a module, it's OK not to spell out the types, and
often helpful because the concrete type may be a mess while the real,
human-facing type is "whatever the caller needs this type to be".
See for example 14fb5e1ce, about a helper within exampleData.