wip sqlite tests: Get a working `immediate` in Jest
TODO should we do something like this in the actual app, too?

The `@expo/websql` package (aka `node-websql`) that provides most of
the JS layer of expo-sqlite uses in turn the `immediate` package to
get an implementation of basically `setImmediate`.

In Jest tests, that doesn't work; the symptom is that the test hangs
and then times out at 5 seconds, not having completed.  Looking at the
`immediate` implementation, it relies on `setTimeout` -- so that'd do it.

We live in a JS world that has Promises.  Given those, setImmediate
has a trivial implementation.  Replace the whole module with that.