persist [nfc]: Cut now-disused "purge" feature
The right layer for something like this to happen at is as a
migration.  We now have a migrations framework that operates at
the right level to be able to do this.  We've used that already to
replace the one way we were using this feature; for any future
occasions where we'd want something like this, the right answer
will again be to make a (new-style) migration.