wip smooth-scroll no polyfill, call explicitly
The methods this was polyfilling all have names starting with
"scroll".  The following finds all the places we were calling them:

  $ git grep -C4 '\.scroll\w*\(' src/webview/js/js.js

Moreover, whenever there wasn't a `behavior: 'smooth'` option,
the polyfill would end up calling the built-in version.  Only one
of these call sites passed such an option.

It appears that Safari doesn't support the "smooth" behavior:
  https://developer.mozilla.org/en-US/docs/Web/API/Window/scroll
So we do still need this polyfill if we want to get that behavior.
... Though whether the polyfill *succeeds* at giving us behavior
we want is another question.