wip pull out inbound events
NB this sends the 'ready' signal slightly earlier: before installing
the press handlers.  That seems fine and appropriate: it needs to come
after we've installed the 'message' event listener, so that we
actually get the events, but doesn't need to interact with whether
we'll notice the user pressing on something.