wip deps: Upgrade react-native-safe-area-context and react-native-screens
TODO manual testing

These need an upgrade in order to upgrade react-navigation to 6.x:
  https://reactnavigation.org/docs/upgrading-from-5.x/#minimum-requirements

Upgrading react-native-safe-area-context calls for updating its type
definitions to match; upstream is written in TypeScript and provides
TS type-definition files.

Ideally this would mean having a script that takes those upstream
type definitions as input, runs something like flowgen with whatever
munging is necessary, and produces suitable output.  As we discover
more rough spots that need munging, we'd express those in some form
in the script (or in flowgen itself) -- in a principled way where
possible, in a hacky way when necessary -- so that rerunning the
script on future slightly-changed versions of the library would
continue to produce suitable output.

Now that we've been moving our type definitions into `.js.flow`
files, and consequently have a much better iteration cycle on them
than we've had before, that may be a direction we move in.  But for
now, for one of our very first upgrades of a library we have
`.js.flow` files for, I did it by hand.

What we have for producing a `.js.flow` file is:
 * Run `yarn flowgen -o /dev/stdout node_modules/…/foo.d.ts`
   to get a rough draft.
 * Fix intersections `&` to be spreads.
 * Fix various React constructs from its TS type definitions to use
   Flow and its Flow type definitions instead.
 * Fix some react-native imports to import from specific modules
   instead.
 * Add our more informative header.

My process was:
 * See what changed upstream:
   * `yarn install` while at old version; `cp -a` the library from
     `node_modules` out to a temporary directory.
   * `yarn install` while at new version; similarly copy the new
     version to a temporary directory.
   * `git diff --no-index` to compare the old and new version.
 * To save work, ignore files where nothing externally-visible
   changed.
 * Where something did change, use the above procedure to make a
   new version.  Use the existing `.js.flow` of the old version
   as a guide for the fixes needed.
 * In one affected file, the existing `.js.flow` had a number of
   minor differences from the flowgen output on the old version.
   I made a prep commit to reconcile those.