wip flowgen-fix: drop rewriting React.RefAttributes; leave that to flowgen
This has the downside that we lose exactness on the object.  (Maybe
find a way to hack that -- like let the rewrite within flowgen add a
comment like `/* flowgen:exact */`, and the printing of object type
respect that?)

It cuts a bunch of `typeof`s.  Hmm, seems like all of them.  Is there
an example where we need a typeof?  flowgen does try to decide whether
one is needed or not, in general.