WIP deps: upgrade react-nav to 6.x
TODO tsflower; read release notes; manual test

TODO branch: squash type updates; repack patch numbering

    $ yarn upgrade --latest -- $(tools/deps | grep @react-navigation/)
    $ yarn add @react-navigation/bottom-tabs

The latter step is to return to a mainline version of
`@react-navigation/bottom-tabs`.  In c0b17bd32 we switched to a
temporary fork, but that was a backport to React Navigation 5.x
of a fix that's already in 6.x.