tools/tsflower: In `check`, handle files removed by patches
Before this change, if a patch removed a file entirely, then after a
successful `tools/tsflower check` the file would be left lying around.
This was because our cleanup at the end relied on `git reset --hard`,
and after the file was restored by unapply_patches it'd be untracked
by Git.

More insidiously, if the TsFlower run actually modified such a file --
that is, if the upstream type definition no longer matched what the
patch was based on, which after all is the main situation that
`tools/tsflower check` exists to check for -- then we wouldn't notice,
again because the file would be untracked by Git and so not considered
by `git diff`.

To fix the issue, after unapplying patches, add new as well as
modified files to the index.

Then to make sure that doesn't end up pulling in any pre-existing
untracked files, which would produce very confusing results and
potentially lead to data loss if the developer wanted those files,
check for them before the command starts work.