offline notice [nfc]: Simplify code for announcements state machine
The behavior here is a machine with three states.  By representing
the state as a single ref with three values, rather than two
booleans, we can make the code a bit more transparent for seeing
that, and also make the three cases shorter and more parallel.

To see that this is NFC, note that in the previous version once
haveAnnouncedUncertain gets set to true, it won't get set to false
again without also setting haveAnnouncedOffline to either true or
false.  So in the condition "haveAnnouncedUncertain is true", it
doesn't matter what value haveAnnouncedOffline has, and that
condition can be analyzed as just one state instead of two.