android build: Bump compileSdkVersion to 33, aka Android 13
This version is out; time to start using it in the build.

This setting is not to be confused with the targetSdkVersion.
The latter goes into the built manifest, and affects a wide
range of behavior, so bumping it requires careful testing.
This is used purely at build time, and should have no effect on
runtime behavior.

Its main effect is that it become possible for code to conditionally
use new API features.  It also brings new compiler warnings --
hence the pair of library updates preceding this.