compose [nfc]: Rename to invocationId from serialNumber
The term "serial number" we use in the translated string here
is a bit of a white lie: a real serial number should always
increase, and never get re-used.

We call it that for the sake of giving translators something
comprehensible without the context of the code around it,
but in the code itself we can use a different term.  This also
aligns with the name `activeInvocations`.