tools/test: Fix name collision on "files", saying "opt_files"
We've had the name "files" for both a global representing the
user's choice of which files to operate on, and a local inside
the run_lint function.

This collision was there for a long time without causing a live
problem, and then started doing so with f3aad7ea2 where we moved
a function call that consulted the global so that it occurred
inside run_lint, with the local active.  (Bash "local" variables
are only sort of local -- they're dynamically scoped, rather than
lexically scoped.)

The effect of the issue is that the "lint" suite would always
succeed without checking anything, because `files_js` would
always produce no output.

Fix it by giving the global a longer, more explicit name.