wip storage: Switch to SQLite's user_version pragma to track migrations
TODO manual tests; issue ref

Recording the version out of band like this, rather than in a table
that's itself part of the schema being tracked, is nice in principle
because it's cleaner.

It's also helpful quite concretely in that Android's `SQLiteDatabase`
library uses this mechanism.  So this will simplify things for us in
using that library when we implement # 5117.