wip copy testing/*.java from RN
This is a crude hack and I'd like to replace it with expressing some
kind of reference in Gradle, rather than making these copies.

But doing this for now so that I can unblock on the Gradle part,
having spent a few hours on that, and explore other parts of what
we need here.

Done like so:

  $ rsync -vaR --exclude 'BUCK' \
      node_modules/react-native/ReactAndroid/./src/androidTest/java/com/facebook/react/testing/ \
      android/app/

plus deleting the previous attempt from the build.gradle.