i18n: Polyfill Intl.PluralRules and Intl.Locale, with Format.JS
These don't currently exist in Hermes:
  https://github.com/facebook/hermes/blob/v0.11.0/doc/IntlAPIs.md

(That doc describes Hermes on Android, but from things the maintainers
have said elsewhere it doesn't sound like the iOS version will be any
more complete.)

We'll need them when using plural formats in translated strings, as we
do for example in ReadReceiptsScreen.  Without them, on visiting that
screen with Hermes we just see the default message (complete with the
pluralization syntax), and the console shows this error (reformatted):

     ERROR  [Error: [`@formatjs/intl` Error FORMAT_ERROR]
      Error formatting message: "{num_of_people, plural, […]}",
      using default message as fallback.

    […]
    Intl.PluralRules is not available in this environment.
    Try polyfilling it using `@formatjs/intl-pluralrules`

These polyfills are automatically conditional, so they won't do
anything while we're not yet on Hermes.