read receipts [nfc]: Silence React warning by providing a "key"
Without this, we get the following warning in the console when
visiting this screen:

   ERROR  Warning: Each child in a list should have a unique "key" prop.

  Check the render method of `FormattedMessage`. See
  https://reactjs.org/link/warning-keys for more information.

In fact, React's default behavior of using the index as the key is
just fine here, because this isn't a list where items can possibly get
inserted, deleted, or reordered over time so that it becomes important
to track their identity.

But it looks like the most reasonable way to get React to stop warning
about this is to explicitly give it those default keys.  So do that.