getHaveServerData: Fix unsafe zulipVersion access causing crash on login
This fixes a regression introduced in 3b811d6af that made it
impossible to log into a new account, by crashing just after
completing authentication with the message "zulipVersion must be
non-null".

Oops: getHaveServerData was not an OK place to call
getServerVersion. It has the following invariant:

  // This invariant will hold as long as we only call this function in a
  // context where we have server data.
  //
  // TODO(#5006): […]
  invariant(zulipVersion, 'zulipVersion must be non-null');

So, abandon that selector and access zulipVersion safely.

While I was at it, I double-checked `Account['zulipVersion']` and
`Account['zulipFeatureLevel']` …and found that their jsdocs weren't
right about when those fields were null. Fix that, copying some text
from the jsdoc of `Account['userId']`.