wipref api: Add "follow" user-topic state, as synonym of "unmuted"
In all the places we currently consult the user's visibility policy
for a topic, treating "follow" as "unmuted" will give it the most
accurate behavior from among what we currently implement.

In particular this makes it so that when the stream is muted, the
topic will correctly be treated as not muted.

There's still more we could do to fully implement this feature:
notably, surface in the UI what the current state is, and give the
option to set a topic to followed.  But this is an important start.