webview [nfc]: Fix readme about how static build works; update list
This README file accurately described how the build worked
when it was written.  But then a few months later we changed that,
in 0fb3d162c, and we didn't update the README accordingly.

The automatic way seems like a nicer developer experience.  Probably
what this script ought to do is to ask Git for the list of files,
and copy those, in order to make it automatic while not getting
fooled by extraneous files in the directory.

For this legacy codebase, though, the way it actually works will
be fine.  Just fix the README to reflect that; and then update
the list to reflect the one change it should have gotten, from
commit ae7ad9965 a few more months later.